requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.2
captcha==0.6.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
confluent-kafka==2.3.0
cryptography==42.0.5
et-xmlfile==1.1.0
flash==1.0.3
future==1.0.0
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
httplib2==0.22.0
humansize==1.0
idna==3.4
importlib-metadata==7.0.0
IPy==1.1
itchat==1.3.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
netmiko==4.3.0
ntc_templates==4.4.0
numpy==1.24.4
openpyxl==3.1.4
pandas==2.0.3
paramiko==3.4.0
pillow==10.4.0
protobuf==4.25.3
psutil==5.9.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyMySQL==1.1.0
PyNaCl==1.5.0
pyparsing==3.1.1
pypng==0.20220715.0
PyQRCode==1.2.1
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
scp==0.14.5
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.25
sqlparse==0.4.4
textfsm==1.1.3
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
Werkzeug==3.0.1
zipp==3.17.0
flask
